+++ backend/src/main/java/SANTA/backend/global/jwt/JWTFilter.java
@@ -1,4 +1,4 @@
-package SANTA.backend.core.auth.jwt;
+package SANTA.backend.global.jwt;
 
 import SANTA.backend.core.user.domain.User;
 import SANTA.backend.core.auth.service.CustomUserDetails;
