+++ backend/src/main/java/SANTA/backend/global/jwt/LoginFilter.java
@@ -1,4 +1,4 @@
-package SANTA.backend.core.auth.jwt;
+package SANTA.backend.global.jwt;
 
 import SANTA.backend.core.auth.dto.LoginRequestDTO;
 import com.fasterxml.jackson.databind.ObjectMapper;
